feat(user): add Password entity to user package
